# Improve where_cond performance for metal